feat: Refactor fonts and themes, optimize imports
This commit removes unused font families from constants and updates the Google Fonts import in `index.html` to reflect these changes. It also cleans up unused components and imports in `App.tsx` and refactors the `AppPreferences` interface in `types.ts` to better structure font and theme-related settings.

The `pt-BR.ts` translation file has also been updated to remove redundant or outdated descriptions for fonts and themes, aligning with the latest structure.

Finally, a minor adjustment is made to `PortfolioContext.tsx` to clarify the purpose of the `setFont` function.